pyproject.toml
[project]
name = "omurtag"
version = "0.2.1"
description = ""
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gitpython>=3.1.46",
    "rich-argparse>=1.7.2",
    "tqdm>=4.67.3",
]

[project.scripts]
omurtag = "omurtag.__main__:main"

[tool.black]
line-length = 80
